feat(goods): power + funding chips on Buyer Pipeline; lift PowerChip to shared
Extends the cross-system power and funding-history overlays onto the Buyer
Pipeline surface (the procurement-wedge cohort that Funder Insight deliberately
excludes). BuyerPipelineRow.id is the goods_relationships.id both overlay maps
key on, so it's the same page-level parallel-fetch + merge-by-rel-id pattern as
Funder Insight / Warm Intros — the existing buyer loader is untouched.

Lifts PowerChip out of insight/page.tsx into a shared _components/goods-power-chip
(FundingChip was already shared) so both surfaces share one definition.

typecheck clean. Chips degrade silently for buyers with no entity match;
coverage grows once the name-only resolution backfill (prior commit) applies.
