+++ apps/web/src/app/org/[slug]/goods/insight/page.tsx
@@ -3,11 +3,12 @@ import { notFound } from 'next/navigation';
 import { ACT_FAST_PROFILE, isActSlug, shouldUseFastLocalOrg } from '@/lib/services/fast-local-org';
 import { getOrgProfileBySlug } from '@/lib/services/org-dashboard-service';
 import { getGoodsFunderInsight } from '@/lib/services/goods-funder-insight';
-import { getGoodsRelationshipPower, powerBand, type RelationshipPower } from '@/lib/services/goods-relationship-power';
+import { getGoodsRelationshipPower, type RelationshipPower } from '@/lib/services/goods-relationship-power';
 import { getGoodsRelationshipFunding, type RelationshipFunding } from '@/lib/services/goods-relationship-funding';
 import { TEMP_LABEL, temperatureTone, type FunderInsight, type InsightFlag } from '@/lib/services/goods-funder-insight-shared';
 import { relDays, money, moneyShort } from '@/lib/services/goods-engagement-shared';
 import { GoodsSubNav } from '../_components/goods-sub-nav';
+import { PowerChip } from '../_components/goods-power-chip';
 import { FundingChip } from '../_components/goods-funding-chip';
 
 export const dynamic = 'force-dynamic';
@@ -94,30 +95,6 @@ function FramingBlock({ framing }: { framing: FunderInsight['framing'] }) {
         <div className="mt-0.5 text-[10px] font-bold text-bauhaus-muted">Contact: {framing.primary_contact}</div>
       )}
     </div>
-  );
-}
-
-/**
- * Cross-system power chip — how embedded this funder is across the 7 power
- * systems, and whether they sit in the revolving-door set. A high-reach yes
- * carries weight and opens doors; it is also a public profile worth knowing
- * before we attach the Goods name. Detail on hover. Renders nothing without signal.
- */
-function PowerChip({ power }: { power: RelationshipPower | null }) {
-  if (!power || power.systemCount < 1) return null;
-  const band = powerBand(power);
-  const tone =
-    band === 'high' ? 'bg-bauhaus-black text-bauhaus-yellow'
-    : band === 'notable' ? 'bg-bauhaus-blue text-white'
-    : 'bg-bauhaus-canvas text-bauhaus-muted';
-  const rd = power.revolvingDoor ? ` · revolving door (${power.vectors.join(', ')})` : '';
-  return (
-    <span
-      className={`px-1.5 py-0.5 text-[9px] font-black uppercase tracking-widest ${tone}`}
-      title={`Cross-system reach: ${power.systems.join(', ') || 'none'}${rd}`}
-    >
-      Power {power.systemCount}/7{power.revolvingDoor ? ' · RD' : ''}
-    </span>
   );
 }
